Extract todo from todos.js into indepent todo.js
