+++ src/reducers/todos.js
@@ -1,26 +1,5 @@
 import { combineReducers } from 'redux';
-
-export const todo = (state, action) => {
-  switch (action.type) {
-  case 'ADD_TODO':
-    return {
-      id: action.id,
-      text: action.text,
-      completed: false
-    };
-  case 'TOGGLE_TODO':
-    if (state.id !== action.id) {
-      return state;
-    }
-
-    return {
-      ...state,
-      completed: !state.completed
-    }
-  default:
-    return state;
-  }
-};
+import todo from './todo';
 
 const byId = (state = {}, action) => {
   switch (action.type) {
